+++ src/main/java/dan200/computercraft/fabric/mixin/HeldItemRendererAccess.java
@@ -1,4 +1,4 @@
-package dan200.computercraft.mixin;
+package dan200.computercraft.fabric.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
