+++ src/main/java/dan200/computercraft/shared/util/IDAssigner.java
@@ -21,7 +21,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import dan200.computercraft.ComputerCraft;
-import dan200.computercraft.mixin.WorldSavePathAccess;
+import dan200.computercraft.fabric.mixin.WorldSavePathAccess;
 import me.shedaniel.cloth.api.utils.v1.GameInstanceUtils;
 
 import net.minecraft.server.MinecraftServer;
