+++ src/main/resources/computercraft.mixins.json
@@ -1,6 +1,6 @@
 {
     "required": true,
-    "package": "dan200.computercraft.mixin",
+    "package": "dan200.computercraft.fabric.mixin",
     "compatibilityLevel": "JAVA_8",
     "mixins": [
         "MinecraftServerAccess",
